Add migración en orden

+++ database/migrations/2026_01_20_040707_change_eps_to_id_eps_in_personas_table.php
@@ -30,8 +30,10 @@ return new class extends Migration
         });
 
         Schema::table('personas', function (Blueprint $table) {
-            $table->dropForeign(['eps_id']);
-            $table->dropColumn('eps_id');
+            if (Schema::hasColumn('personas', 'eps_id')) {
+                $table->dropForeign(['eps_id']);
+                $table->dropColumn('eps_id');
+            }
         });
     }
 };
